OLED Update & Upgrade
I've added OLED capability for live-time gear change readout. Provided with the U8g2 library and I2C. There is still serial output with voltage values and I will continue to use this capability. Board prototyping is next. I'm leaving it up to you if you want to change the fonts - OLED screens are not very forgiving when it comes to adjusting X,Y values along with fonts!!!